// src/api/generated/schemas/deploymentGroup.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * NervesCloud API
 * The NervesCloud API gives users full access to their
Orgs, Products, and corresponding Device fleets.

The API can be used to integrate with your own systems, providing full access to your Product and Device data.

The API is documented using the OpenAPI 3.0 specification.

 * OpenAPI spec version: 2.0.0
 */
import type { Conditions } from './conditions';
import type { CurrentRelease } from './currentRelease';
import type { DeploymentGroupState } from './deploymentGroupState';

export interface DeploymentGroup {
  conditions?: Conditions;
  current_release?: CurrentRelease;
  delta_updatable?: boolean;
  device_count?: number;
  firmware_uuid?: string;
  is_active?: boolean;
  name?: string;
  releases_count?: number;
  state?: DeploymentGroupState;
  /** @nullable */
  id?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  archive_uuid?: string | null;
}